pyproject.toml
[project]
name = "rave-tcv"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "acids-rave>=2.3.1",
    "ffmpeg>=1.4",
    "gin>=0.1.6",
    "librosa>=0.11.0",
    "matplotlib>=3.10.3",
    "music2latent>=0.1.6",
    "ruff>=0.12.9",
    "torchaudio>=2.5.0",
    "tqdm>=4.67.1",
]
